Initial separation of fetchers logic
