+++ src/eunomia/fetchers/__init__.py
@@ -1,0 +1,3 @@
+from eunomia.fetchers.internal import EunomiaInternalFetcher
+
+__all__ = ["EunomiaInternalFetcher"]
